Change HIDPowerDevice_::sendDate "id" argument from 16 to 8bits.

+++ src/HIDPowerDevice.cpp
@@ -233,7 +233,7 @@ HIDPowerDevice_::HIDPowerDevice_(void) {
     SetFeature(HID_PD_MANUFACTURER, &bManufacturer, sizeof(bManufacturer));
 }
 
-int HIDPowerDevice_::sendDate(uint16_t id, uint16_t year, uint8_t month, uint8_t day) {
+int HIDPowerDevice_::sendDate(uint8_t id, uint16_t year, uint8_t month, uint8_t day) {
     uint16_t bval = (year - 1980)*512 + month * 32 + day;
     return SendReport(id, &bval, sizeof (bval));
 }
